Tests work with reload and gulp now
