+++ README.md
@@ -30,6 +30,9 @@ yarn build
 
 # run the packer script targeted for 'chrome' or 'firefox' after build
 yarn bundle
+
+# build the tests
+yarn build-tests
 ```
 
 For detailed explanation on how things work, consult the [docs for vue-loader](http://vuejs.github.io/vue-loader).
