+++ services/keepassReference.js
@@ -73,9 +73,16 @@ function KeepassReference() {
 			}
 		}
 
+		// https://stackoverflow.com/questions/2970525/converting-any-string-into-camel-case
+		let camelize = (str) => {
+			return str.replace(/(?:^\w|[A-Z]|\b\w)/g, function(letter, index) {
+				return index == 0 ? letter.toLowerCase() : letter.toUpperCase();
+			}).replace(/\s+/g, '');
+		}
+
 		var customLocalString = /^\{S:([a-zA-Z]+)\}$/.exec(referenceText)
 		if (customLocalString) {
-			var camelCase = Case.camel(customLocalString[1])
+			var camelCase = camelize(customLocalString[1])
 			return currentEntry[camelCase];
 		}
