+++ tests/services/test.protectedMemory.js
