+++ tests/test.html
@@ -1,15 +1,6 @@
 <html>
-<head>
-  <meta charset="utf-8">
-  <title>Mocha Tests - All</title>
-  <link rel="stylesheet" href="../node_modules/mocha/mocha.css" />
-</head>
 <body>
-  <div id="mocha"></div>
-  <script type="text/javascript" src="bundle.js"></script>
-  <script>
-    mocha.checkLeaks();
-    mocha.run();
-  </script>
+	<!-- Mocha loader sets everything up for us -->
+  <script type="text/javascript" src="bundle.build.js"></script>
 </body>
 </html>
