+++ tests/test.js
@@ -1,25 +1,3 @@
-/*
-  <script src="../node_modules/mocha/mocha.js"></script>
-  <script>mocha.setup('bdd')</script>
-  <script src="../node_modules/should/should.js"></script>
-  <script src="../node_modules/sinon/pkg/sinon.js"></script>
-  <script src="../lib/base64.js"></script>
-  <script src="../lib/angular.min.js"></script>
-  <script src="../lib/case.min.js"></script>
-  <script src="../lib/salsa20.js"></script>
-  <script src="../bower_components/kdbxweb/dist/kdbxweb.js"></script>
-  <script src="../bower_components/angular-mocks/angular-mocks.js"></script>
-  <script src="../services/settings.js"></script>
-  <script src="../services/protectedMemory.js"></script>
-  <script src="../services/secureCacheMemory.js"></script>
-  <script src="../services/secureCacheDisk.js"></script>
-  <script src="../services/oneDriveFileManager.js"></script>
-  <script src="../services/keepassReference.js"></script>
-  <script src="test.secureCache.js"></script>
-  <script src="test.protectedMemory.js"></script>
-  <script src="test.oneDriveFileManager.js"></script>
-  <script src="test.keepassReference.js"></script>
-*/
-require('./test.keepassReference.js')
-require('./test.protectedMemory.js')
-require('./test.secureCache.js')
+require('./services/test.keepassReference.js')
+require('./services/test.protectedMemory.js')
+require('./services/test.secureCache.js')
